refactor: extract vim types into types.ts
